refactor: merge storage call executor into storage
